+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- feat: Touch events take Regex for ignoreNames & add tests #973
+
 ## 1.6.2
 
 - fix: Don't prefix app:/// to "native" filename as well #957